+++ cli.js
@@ -3,7 +3,7 @@ let deploy = require('.')
 let { banner, updater } = require('@architect/utils')
 let create = require('@architect/create')
 let validate = require('./src/validate')
-let options = require('./src/options')
+let _options = require('./src/options')
 let { version } = require('./package.json')
 let pauser = require('./src/utils/pause-sandbox')
 
@@ -20,10 +20,14 @@ let pauser = require('./src/utils/pause-sandbox')
  * -t|--tags|tags ............... add tags
  * -n|--name|name ............... customize stack name
  */
-async function cmd (opts = []) {
+async function cmd () {
+  let opts = process.argv || []
 
   // Validate for expected env and args and check for potential creds issues
-  validate(opts)
+  validate()
+
+  // Populate options, read args into `prune`, `verbose`, `production`, `tags`, `name`, `isFullDeploy`, etc.
+  let options = _options(opts)
 
   // Pause the Sandbox watcher
   pauser.pause()
@@ -31,23 +35,20 @@ async function cmd (opts = []) {
   // create any missing local infra
   await create({})
 
-  // read args into {prune, verbose, production, tags, name, isFullDeploy}
-  let args = options(opts)
-
-  if (args.isDirty || args.srcDirs.length) {
-    let result = await deploy.dirty(args)
+  if (options.isDirty || options.srcDirs.length) {
+    let result = await deploy.dirty(options)
     pauser.unpause()
     return result
   }
 
-  if (args.isStatic) {
-    let result = await deploy.static(args)
+  if (options.isStatic) {
+    let result = await deploy.static(options)
     pauser.unpause()
     return result
   }
 
-  // deploy with SAM by default..
-  let result = await deploy.sam(args)
+  // Deploy with SAM by default..
+  let result = await deploy.sam(options)
   pauser.unpause()
   return result
 }
@@ -56,19 +57,16 @@ module.exports = cmd
 
 // Allow direct invoke
 if (require.main === module) {
-  let args = options(process.argv)
-  ;(async function () {
+  (async function () {
     try {
       banner({ version: `Deploy ${version}` })
-      await cmd(process.argv)
+      await cmd()
     }
     catch (err) {
       // Unpause the Sandbox watcher
       pauser.unpause()
       let update = updater('Deploy')
       update.error(err)
-      if (args.verbose)
-        console.log(err)
     }
   })()
 }