+++ src/dirty/deploy-one.js
@@ -1,7 +1,4 @@
 let series = require('run-series')
-let parse = require('@architect/parser')
-let fs = require('fs')
-let path = require('path')
 let parallel = require('run-parallel')
 let sploot = require('run-waterfall')
 let zip = require('./zip')
@@ -13,21 +10,15 @@ let aws = require('aws-sdk')
  * ...as quickly as possible
  *
  * @param {String} params.FunctionName - a valid lambda function name or arn
- * @param {String} params.pathToCode - path on filesystem to function code
+ * @param {String} params.lambda - Inventory Lambda object
  */
-module.exports = function dirtyDeployOne ({ FunctionName, pathToCode }, callback) {
+module.exports = function dirtyDeployOne (params, callback) {
   parallel({
     code (callback) {
-      updateCode({
-        FunctionName,
-        pathToCode,
-      }, callback)
+      updateCode(params, callback)
     },
     config (callback) {
-      updateConfig({
-        FunctionName,
-        pathToCode,
-      }, callback)
+      updateConfig(params, callback)
     }
   },
   function done (err) {
@@ -39,10 +30,11 @@ module.exports = function dirtyDeployOne ({ FunctionName, pathToCode }, callback
 /**
  * zip path/to/code and dirty updates lambda function
  */
-function updateCode ({ FunctionName, pathToCode }, callback) {
+function updateCode ({ FunctionName, lambda }, callback) {
+  let { src } = lambda
   sploot([
     function (callback) {
-      zip(pathToCode, callback)
+      zip(src, callback)
     },
     function (buffer, callback) {
       let lambda = new aws.Lambda({ region: process.env.AWS_REGION })
@@ -61,63 +53,48 @@ function updateCode ({ FunctionName, pathToCode }, callback) {
 /**
  * reads path/to/code/.arc-config and dirty updates lambda function config
  */
-function updateConfig ({ FunctionName, pathToCode }, callback) {
+function updateConfig (params, callback) {
+  let { FunctionName } = params
+  let { timeout, memory, runtime, handler, concurrency, layers, policies } = params.lambda.config
+
   let lambda = new aws.Lambda({ region: process.env.AWS_REGION })
-  let configPath = path.join(pathToCode, '.arc-config')
-  if (fs.existsSync(configPath)) {
-    fs.readFile(configPath, 'utf8', function done (err, body) {
-      if (err) callback()
-      else {
-        let configuration = parse(body.toString())
-        if (!configuration.aws) {
-          callback()
-        }
-        else {
-          // turn {aws:[['timeout', 3], ['runtime', 'ruby2.5']]
-          // into {timeout:3, runtime:'ruby2.5'}
-          let pivot = configuration.aws.reduce((a, tuple) => {
-            a[tuple[0]] = tuple[1]
-            return a
-          }, {})
-          // override function configuration
-          let params = { FunctionName }
-          if (pivot.memory)
-            params.MemorySize = pivot.memory
-          if (pivot.timeout)
-            params.Timeout = pivot.timeout
-          if (pivot.runtime)
-            params.Runtime = pivot.runtime
-          if (pivot.layers)
-            params.Layers = pivot.layers
-          series([
-            function updateFunctionConfiguration (callback) {
-              setTimeout(function rateLimit () {
-                lambda.updateFunctionConfiguration(params, callback)
-              }, 200)
-            },
-            function updateFunctionConcurrency (callback) {
-              if (!pivot.concurrency || pivot.concurrency === 'unthrottled') {
-                setTimeout(function rateLimit () {
-                  lambda.deleteFunctionConcurrency({
-                    FunctionName,
-                  }, callback)
-                }, 200)
-              }
-              else {
-                setTimeout(function rateLimit () {
-                  lambda.putFunctionConcurrency({
-                    FunctionName,
-                    ReservedConcurrentExecutions: pivot.concurrency,
-                  }, callback)
-                }, 200)
-              }
-            }
-          ], callback)
-        }
+
+  let args = {
+    FunctionName,
+    Handler: handler,
+    MemorySize: memory,
+    Timeout: timeout,
+    Runtime: runtime,
+    // TODO addme! (need stage)
+    // Environment: {
+    //   Variables: {}
+    // },
+  }
+  if (layers.length > 0) args.Layers = layers
+  if (policies.length > 0) args.Policies = policies
+
+  series([
+    function updateFunctionConfiguration (callback) {
+      setTimeout(function rateLimit () {
+        lambda.updateFunctionConfiguration(args, callback)
+      }, 200)
+    },
+    function updateFunctionConcurrency (callback) {
+      if (!concurrency || concurrency === 'unthrottled') {
+        setTimeout(function rateLimit () {
+          lambda.deleteFunctionConcurrency({
+            FunctionName,
+          }, callback)
+        }, 200)
       }
-    })
-  }
-  else {
-    callback()
-  }
+      else {
+        setTimeout(function rateLimit () {
+          lambda.putFunctionConcurrency({
+            FunctionName,
+            ReservedConcurrentExecutions: concurrency,
+          }, callback)
+        }, 200)
+      }
+    }
+  ], callback)
 }