+++ src/dirty/index.js
@@ -1,8 +1,6 @@
-let pkg = require('@architect/package')
-let utils = require('@architect/utils')
-let updater = utils.updater
-let deployNested = require('./deploy-nested-stack')
-let deployStack = require('./deploy-sam-stack')
+let { toLogicalID, updater } = require('@architect/utils')
+let { sep } = require('path')
+let deploySAM = require('./deploy-sam')
 let pretty = require('./pretty')
 
 /**
@@ -12,60 +10,49 @@ let pretty = require('./pretty')
  * @param {Function} callback - node style errback
  * @returns {Promise} if no callback is supplied
  */
-module.exports = function dirty ({ isDryRun = false, srcDirs = [] }, callback) {
-  // return a promise if a callback is not supplied
-  let promise
-  if (!callback) {
-    promise = new Promise(function ugh (res, rej) {
-      callback = function errback (err, result) {
-        if (err) rej(update.err(err))
-        else res(result)
-      }
-    })
-  }
+module.exports = function dirty (inventory, params, callback) {
+  let { isDryRun = false, srcDirs = [] } = params
+  let { inv } = inventory
 
   // update console output
   let update = updater('Deploy')
+  if (isDryRun) {
+    update = updater('Deploy [dry-run]')
+    update.status('Starting dry run!')
+  }
+
+  // time the deploy
+  let ts = Date.now()
+  let specificLambdasToDeploy = []
+  if (srcDirs.length && inv.lambdaSrcDirs.length) {
+    // Strip trailing slashes
+    srcDirs = srcDirs.map(d => d.endsWith(sep) ? d.substr(0, d.length - 1) : d)
+    specificLambdasToDeploy = srcDirs.filter(d => inv.lambdaSrcDirs.some(p => p.endsWith(d)))
+  }
+
+  let appname = inv.app
+  let stackname = `${toLogicalID(appname)}Staging`
+
+  update.warn('Direct deployments should be considered temporary, and will be overwritten')
+  update.status(
+    'Initializing direct deployment',
+    `Stack ... ${stackname}`
+  )
+  pretty.warn(update)
 
   if (isDryRun) {
-    // TODO implement dry run?
-    update.status('Dirty dry run not yet available, skipping dirty deploy...')
+    let plural = srcDirs.length > 1 ? 's' : ''
+    update.status(`Direct deploy to Lambda${plural}:`, ...srcDirs)
     callback()
   }
+
   else {
-    // time the deploy
-    let ts = Date.now()
-    let inventory = utils.inventory()
-    let specificLambdasToDeploy = []
-    if (srcDirs.length && inventory.localPaths.length) {
-      specificLambdasToDeploy = srcDirs.reduce((acc, d) => acc.concat(inventory.localPaths.filter(p => p.startsWith(d))), [])
-    }
-    let { arc } = utils.readArc()
-
-    // FIXME architect/package is mutating the orig arc object and adding a (possibly) non existent get / to http
-    let copy = JSON.parse(JSON.stringify(arc))
-
-    let appname = arc.app[0]
-    let stackname = `${utils.toLogicalID(appname)}Staging`
-    let sam = pkg(arc)
-    let nested = Object.prototype.hasOwnProperty.call(sam, `${appname}-cfn.json`)
-    let exec = nested ? deployNested : deployStack
-
-    update.warn('Direct deployments should be considered temporary, and will be overwritten')
-    update.status(
-      'Initializing direct deployment',
-      `Stack ... ${stackname}`
-    )
-    pretty.warn(update)
-
-    exec({
-      ts,
+    deploySAM({
+      inventory,
       specificLambdasToDeploy,
-      arc: copy,
       stackname,
+      ts,
       update
     }, callback)
   }
-
-  return promise
 }