+++ src/sam/index.js
@@ -1,8 +1,6 @@
 let pkg = require('@architect/package')
-let utils = require('@architect/utils')
-let { readArc, toLogicalID, updater } = require('@architect/utils')
-let fingerprinter = utils.fingerprint
-let fingerprintConfig = fingerprinter.config
+let { toLogicalID, updater, fingerprint: fingerprinter } = require('@architect/utils')
+let { config: fingerprintConfig } = fingerprinter
 let series = require('run-series')
 let hydrate = require('@architect/hydrate')
 
@@ -21,16 +19,24 @@ let after = require('./02-after')
  * @param {Function} callback - a node-style errback
  * @returns {Promise} - if not callback is supplied
  */
-module.exports = function samDeploy (params, callback) {
-  let { apiType, isDryRun = false, name, production, prune, tags, verbose } = params
+module.exports = function samDeploy (inventory, params, callback) {
+  let {
+    apiType,
+    isDryRun = false,
+    name,
+    production,
+    prune,
+    tags,
+    verbose,
+  } = params
+  let { inv } = inventory
 
   let stage = production ? 'production' : 'staging'
   let ts = Date.now()
   let log = true
   let pretty = print({ log, verbose })
-  let { arc } = readArc()
-  let bucket // Assigned later
-  let appname = arc.app[0]
+  let appname = inv.app
+  let bucket = inv.aws.bucket
   let stackname = `${toLogicalID(appname)}${production ? 'Production' : 'Staging'}`
   let update = updater('Deploy')
 
@@ -41,21 +47,11 @@ module.exports = function samDeploy (params, callback) {
   // Assigned below
   let cloudformation
   let sam
-  let nested
   let foundLegacyApi
 
   let region = process.env.AWS_REGION
-  if (!region)
+  if (!region) {
     throw ReferenceError('AWS region must be configured to deploy')
-
-  let promise
-  if (!callback) {
-    promise = new Promise(function ugh (res, rej) {
-      callback = function errback (err, result) {
-        if (err) rej(err)
-        else res(result)
-      }
-    })
   }
 
   if (isDryRun) {
@@ -64,8 +60,7 @@ module.exports = function samDeploy (params, callback) {
   }
 
   // API switching
-  let findApiType = s => s[0] && s[0] === 'apigateway' && s[1]
-  let arcApiType = arc.aws && arc.aws.some(findApiType) && arc.aws.find(findApiType)[1]
+  let arcApiType = inv.aws.apigateway
 
   series([
     /**
@@ -77,13 +72,8 @@ module.exports = function samDeploy (params, callback) {
         callback()
       }
       else {
-        let bucketProvided = arc.aws && arc.aws.some(o => o[0] === 'bucket')
-        if (bucketProvided) {
-          bucket = arc.aws.find(o => o[0] === 'bucket')[1]
-          callback()
-        }
+        if (bucket) callback()
         else {
-          let appname = arc.app[0]
           getBucket({
             appname,
             region,
@@ -116,7 +106,7 @@ module.exports = function samDeploy (params, callback) {
      * Maybe write static asset manifest prior to cfn or hydration
      */
     function maybeFingerprint (callback) {
-      let { fingerprint } = fingerprintConfig(arc)
+      let { fingerprint } = fingerprintConfig(inv._project.arc)
 
       if (fingerprint || verbose)
         update.done(`Static asset fingerpringing ${fingerprint ? 'enabled' : 'disabled'}`)
@@ -124,12 +114,8 @@ module.exports = function samDeploy (params, callback) {
       // Always run full fingerprinting op to ensure remnant static.json files are deleted
       // This is especially important in Arc 6+ where we no longer do .arc checks for fingerprint status
       fingerprinter({}, function done (err) {
-        if (err) {
-          callback(err)
-        }
-        else {
-          callback()
-        }
+        if (err) callback(err)
+        else callback()
       })
     },
 
@@ -144,7 +130,7 @@ module.exports = function samDeploy (params, callback) {
      * Generate cfn, which must be completed only after fingerprinting or files may not be present
      */
     function generateCloudFormation (callback) {
-      cloudformation = pkg(arc)
+      cloudformation = pkg(inventory)
       callback()
     },
 
@@ -156,7 +142,7 @@ module.exports = function samDeploy (params, callback) {
       if (arcApiType) callback()
       else {
         compat({
-          arc,
+          inv,
           stackname
         }, function done (err, result = {}) {
           if (err) callback(err)
@@ -183,11 +169,8 @@ module.exports = function samDeploy (params, callback) {
       else {
         apiType = 'http'
       }
-
-      if (apiType) {
-        let valid = [ 'http', 'httpv1', 'httpv2', 'rest' ]
-        if (!valid.some(v => v === apiType)) throw ReferenceError(`API type must be 'http[v1|v2]', or 'rest'`)
-      }
+      // Special workflow-specific case where we'll additively mutate the inventory object
+      inv._deploy = { apiType }
       callback()
     },
 
@@ -195,17 +178,14 @@ module.exports = function samDeploy (params, callback) {
      * Macros (both built-in + user)
      */
     function runMacros (callback) {
-      let options = { apiType }
       macros(
-        arc,
+        inventory,
         cloudformation,
         stage,
-        options,
         function done (err, _sam) {
           if (err) callback(err)
           else {
             sam = _sam
-            nested = Object.prototype.hasOwnProperty.call(sam, `${appname}-cfn.json`)
             callback()
           }
         })
@@ -215,7 +195,7 @@ module.exports = function samDeploy (params, callback) {
      * Pre-deploy ops
      */
     function beforeDeploy (callback) {
-      let params = { sam, nested, bucket, pretty, update, isDryRun }
+      let params = { sam, bucket, pretty, update, isDryRun }
       before(params, callback)
     },
 
@@ -231,7 +211,6 @@ module.exports = function samDeploy (params, callback) {
         deploy({
           appname,
           stackname,
-          nested,
           bucket,
           pretty,
           region,
@@ -253,8 +232,7 @@ module.exports = function samDeploy (params, callback) {
       else {
         let legacyAPI = apiType === 'rest'
         let params = {
-          appname,
-          arc,
+          inventory,
           legacyAPI,
           pretty,
           production,
@@ -270,6 +248,4 @@ module.exports = function samDeploy (params, callback) {
     }
 
   ], callback)
-
-  return promise
 }