+++ src/sam/macros/_legacy-api/index.js
@@ -6,8 +6,8 @@ let unexpress = require('./un-express-route')
 let forceStatic = require('./add-static-proxy')
 
 // eslint-disable-next-line
-module.exports = async function legacyAPI (arc, cloudformation, stage, options) {
-  let { apiType } = options
+module.exports = async function legacyAPI (arc, cloudformation, stage, inventory) {
+  let { apiType } = inventory.inv._deploy
   if (apiType === 'rest' && arc.http.length) {
     // Copy arc.http to avoid get index mutation
     let http = JSON.parse(JSON.stringify(arc.http))