+++ src/static/index.js
@@ -2,9 +2,10 @@ let aws = require('aws-sdk')
 let { join } = require('path')
 let { existsSync } = require('fs')
 let waterfall = require('run-waterfall')
-let { readArc } = require('@architect/parser')
 let { fingerprint: fingerprinter, toLogicalID, updater } = require('@architect/utils')
+let { config: fingerprintConfig } = fingerprinter
 let publish = require('./publish')
+let getResources = require('../utils/get-cfn-resources')
 
 /**
  * Upload files to CFN defined bucket
@@ -13,7 +14,7 @@ let publish = require('./publish')
  * @param {Function} callback - a node-style errback
  * @returns {Promise} - if no callback is supplied
  */
-module.exports = function deployStatic (params, callback) {
+module.exports = function deployStatic (inventory, params, callback) {
   let {
     bucket: Bucket,
     credentials,
@@ -31,21 +32,6 @@ module.exports = function deployStatic (params, callback) {
   } = params
   if (!update) update = updater('Deploy')
 
-  // AWS config
-  region = region || process.env.AWS_REGION
-  let config = { region }
-  if (credentials) config.credentials = credentials
-
-  let promise
-  if (!callback) {
-    promise = new Promise(function ugh (res, rej) {
-      callback = function errback (err, result) {
-        if (err) rej(err)
-        else res(result)
-      }
-    })
-  }
-
   if (isDryRun) {
     // TODO implement static deploy dry run?
     update.status('Static dry run not yet available, skipping static deploy...')
@@ -53,68 +39,47 @@ module.exports = function deployStatic (params, callback) {
   }
   else {
     update.status('Deploying static assets...')
-    let { arc } = readArc()
-    let appname = arc.app[0]
+    let { inv } = inventory
+    let appname = inv.app
 
     if (!stackname) {
       stackname = `${toLogicalID(appname)}${production ? 'Production' : 'Staging'}`
       if (name) stackname += toLogicalID(name)
     }
 
-    let staticEnabled = arc.static || (arc.http && existsSync(join(process.cwd(), 'public')))
-
     waterfall([
       // Parse settings
       function (callback) {
         // Bail early if this project doesn't have @static specified
-        if (!staticEnabled) {
-          callback(Error('cancel'))
-        }
+        if (!inv.static) callback(Error('cancel'))
         else {
-          function setting (name, bool) {
-            let value
-            if (!arc.static) return false
-            for (let opt of arc.static) {
-              if (!opt[0]) continue
-              if (opt[0].toLowerCase() === name && opt[1]) {
-                if (bool && opt[1] === true) value = true
-                else value = opt[1]
-              }
-            }
-            return value || false
-          }
-
           // Fingerprinting + ignore any specified files
-          let { fingerprint, ignore } = fingerprinter.config(arc)
+          let { fingerprint, ignore } = fingerprintConfig(inv._project.arc)
 
           // Asset pruning: delete files not present in public/ folder
-          prune = prune || setting('prune', true)
+          prune = prune || inv.static.prune
 
           // Project folder remap
-          let folder = setting('folder') || 'public'
+          let folder = inv.static.folder
           if (!existsSync(join(process.cwd(), folder))) {
             callback(Error('@static folder not found'))
           }
-
-          // Published path prefixing
-          prefix = prefix || setting('prefix')
-
-          callback(null, { fingerprint, ignore, folder })
+          else {
+            // Published path prefixing
+            prefix = prefix || inv.static.prefix
+            callback(null, { fingerprint, ignore, folder })
+          }
         }
       },
 
       // Get the bucket PhysicalResourceId if not supplied
       function (params, callback) {
         if (!Bucket) {
-          let cloudformation = new aws.CloudFormation(config)
-          cloudformation.listStackResources({
-            StackName: stackname
-          },
-          function done (err, data) {
+          getResources({ credentials, region, stackname }, function (err, resources) {
             if (err) callback(err)
             else {
               let find = i => i.ResourceType === 'AWS::S3::Bucket' && i.LogicalResourceId === 'StaticBucket'
-              Bucket = data.StackResourceSummaries.find(find).PhysicalResourceId
+              Bucket = resources.find(find).PhysicalResourceId
               callback(null, params)
             }
           })
@@ -123,6 +88,8 @@ module.exports = function deployStatic (params, callback) {
       },
 
       function ({ fingerprint, ignore, folder }, callback) {
+        let config = { region: region || process.env.AWS_REGION }
+        if (credentials) config.credentials = credentials
         let s3 = new aws.S3(config)
 
         publish({
@@ -146,6 +113,4 @@ module.exports = function deployStatic (params, callback) {
       else callback()
     })
   }
-
-  return promise
 }