+++ src/validate/index.js
@@ -1,7 +1,6 @@
 let { updater } = require('@architect/utils')
-let path = require('path')
-let fs = require('fs')
-let update = updater('Deploy')
+let { delimiter, join } = require('path')
+let { existsSync } = require('fs')
 
 let messages = {
   missing_aws_cli: 'missing aws from PATH, please install the aws-cli',
@@ -9,18 +8,25 @@ let messages = {
   missing_region: '@aws region / AWS_REGION must be configured',
 }
 
-module.exports = function validate (/* opts*/) {
+/**
+ * Deploy CLI environment validator
+ */
+module.exports = function validate () {
   try {
-    if (process.env.ARC_AWS_CREDS === 'missing')
+    if (process.env.ARC_AWS_CREDS === 'missing') {
       throw Error('missing_creds')
+    }
 
-    if (!binExists('aws'))
+    if (!binExists('aws')) {
       throw ReferenceError('missing_aws_cli')
+    }
 
-    if (!process.env.AWS_REGION)
+    if (!process.env.AWS_REGION) {
       throw Error('missing_region')
+    }
   }
   catch (e) {
+    let update = updater('Deploy')
     update.error(`Failed to deploy, ${messages[e.message]}`)
     process.exit(1)
   }
@@ -30,13 +36,13 @@ function binExists (bin) {
   function getPaths (bin) {
     var envPath = (process.env.PATH || '')
     var envExt = (process.env.PATHEXT || '')
-    return envPath.replace(/["]+/g, '').split(path.delimiter).map(function (chunk) {
-      return envExt.split(path.delimiter).map(function (ext) {
-        return path.join(chunk, bin + ext)
+    return envPath.replace(/["]+/g, '').split(delimiter).map(function (chunk) {
+      return envExt.split(delimiter).map(function (ext) {
+        return join(chunk, bin + ext)
       })
     }).reduce(function (a, b) {
       return a.concat(b)
     })
   }
-  return getPaths(bin).some(fs.existsSync)
+  return getPaths(bin).some(existsSync)
 }