+++ test/integration/static/index-test.js
@@ -1,8 +1,14 @@
 let test = require('tape')
 let { join } = require('path')
 let proxyquire = require('proxyquire')
+let inventory = require('@architect/inventory')
 let { updater } = require('@architect/utils')
 let mockFs = require('mock-fs')
+
+// Necessary to run test solo
+let aws = require('aws-sdk')
+let awsMock = require('aws-sdk-mock')
+new aws.S3()
 
 let published
 function publish (params, callback) {
@@ -10,9 +16,9 @@ function publish (params, callback) {
   callback(null, params)
 }
 
-let filePath = join(process.cwd(), 'src', 'static', 'index.js')
-let sut = proxyquire(filePath, {
-  './publish': publish,
+let staticDeployPath = join(process.cwd(), 'src', 'static', 'index.js')
+let staticDeployMod = proxyquire(staticDeployPath, {
+  './publish': publish
 })
 
 let defaultParams = () => ({
@@ -35,15 +41,33 @@ function setup () {
   published = undefined
 }
 function reset () {
+  params = defaultParams()
   mockFs.restore()
 }
 
-// Note: it'd be nice to test the CloudFormation stackname code path
-// However, mock-fs doesn't play nicely with aws-sdk(-mock)
+function staticDeploy (t, callback) {
+  inventory({}, function (err, result) {
+    if (err) t.fail(err)
+    else {
+      staticDeployMod(result, params, err => {
+        reset() // Must be reset before any tape tests are resolved because mock-fs#201
+        callback(err)
+      })
+    }
+  })
+}
+
+/**
+ * Notes:
+ * - Unfortunately, proxyquire seems to have a nested file folder + `@global` bug, so we can't run this from index
+ *   - Instead, we have to run inventory ourselves on each test, which kinda sucks
+ * - Also, it'd be nice to test the CloudFormation stackname code path
+ *   - However, mock-fs doesn't play nicely with aws-sdk(-mock)
+ */
 
 test('Module is present', t => {
   t.plan(1)
-  t.ok(sut, 'Publish module is present')
+  t.ok(staticDeployMod, 'Static asset deployment module is present')
 })
 
 test(`Skip static deploy if @static isn't defined`, t => {
@@ -51,22 +75,23 @@ test(`Skip static deploy if @static isn't defined`, t => {
   setup()
   let arc = '@app\n an-app'
   mockFs({ 'app.arc': arc })
-  sut(params, err => {
-    reset() // Must be reset before any tape tests are resolved because mock-fs#201
+  staticDeploy(t, err => {
     if (err) t.fail(err)
     t.notOk(published, 'Publish not called')
   })
 })
 
-test(`Skip static deploy if @http is defined, but public/ folder is not present`, t => {
-  t.plan(1)
+test(`Static deploy error if @http is defined, but public/ folder is not present`, t => {
+  t.plan(2)
   setup()
   let arc = '@app\n an-app\n @http'
   mockFs({ 'app.arc': arc })
-  sut(params, err => {
-    reset() // Must be reset before any tape tests are resolved because mock-fs#201
-    if (err) t.fail(err)
-    t.notOk(published, 'Publish not called')
+  staticDeploy(t, err => {
+    if (err) {
+      t.ok(err.message.includes('@static folder not found'))
+      t.notOk(published, 'Publish not called')
+    }
+    else t.fail('Did not find error')
   })
 })
 
@@ -78,16 +103,15 @@ test(`Publish static deploy if @static is defined`, t => {
     'app.arc': arc,
     'public': {}
   })
-  sut(params, err => {
-    reset() // Must be reset before any tape tests are resolved because mock-fs#201
+  staticDeploy(t, err => {
     if (err) t.fail(err)
     t.equal(published.Bucket, params.bucket, 'Bucket is unchanged')
     t.equal(published.fingerprint, false, 'Fingerprint set to false by default')
     t.equal(published.folder, 'public', 'Folder set to public by default')
     t.equal(published.ignore.length, 0, 'Ignore is empty by default')
     t.equal(published.isFullDeploy, params.isFullDeploy, 'isFullDeploy is unchaged')
-    t.equal(published.prefix, false, 'Prefix set to false by default')
-    t.equal(published.prune, params.prune, 'Prune is unchaged')
+    t.equal(published.prefix, null, 'Prefix set to null by default')
+    t.equal(published.prune, null, 'Prune set to null by default')
     t.equal(published.region, params.region, 'Region is unchaged')
   })
 })
@@ -100,8 +124,7 @@ test(`Publish static deploy if @http is defined and public/ folder is present`, 
     'app.arc': arc,
     'public': {}
   })
-  sut(params, err => {
-    reset() // Must be reset before any tape tests are resolved because mock-fs#201
+  staticDeploy(t, err => {
     if (err) t.fail(err)
     t.ok(published, 'Publish was called')
   })
@@ -115,8 +138,7 @@ test(`Respect @static fingerprint true`, t => {
     'app.arc': arc,
     'public': {}
   })
-  sut(params, err => {
-    reset() // Must be reset before any tape tests are resolved because mock-fs#201
+  staticDeploy(t, err => {
     if (err) t.fail(err)
     t.equal(published.fingerprint, true, 'Fingerprint set to true')
   })
@@ -130,8 +152,7 @@ test(`Respect @static fingerprint external`, t => {
     'app.arc': arc,
     'public': {}
   })
-  sut(params, err => {
-    reset() // Must be reset before any tape tests are resolved because mock-fs#201
+  staticDeploy(t, err => {
     if (err) t.fail(err)
     t.equal(published.fingerprint, 'external', 'Fingerprint set to external')
   })
@@ -152,8 +173,7 @@ ignore
     'app.arc': arc,
     'public': {}
   })
-  sut(params, err => {
-    reset() // Must be reset before any tape tests are resolved because mock-fs#201
+  staticDeploy(t, err => {
     if (err) t.fail(err)
     t.equal(published.ignore[0], 'foo', 'Got correct ignore config')
     t.equal(published.ignore[1], 'bar', 'Got correct ignore config')
@@ -168,10 +188,8 @@ test(`Respect prune param`, t => {
     'app.arc': arc,
     'public': {}
   })
-  let params = defaultParams()
   params.prune = true
-  sut(params, err => {
-    reset() // Must be reset before any tape tests are resolved because mock-fs#201
+  staticDeploy(t, err => {
     if (err) t.fail(err)
     t.ok(published.prune, 'Prune is unchaged')
   })
@@ -185,8 +203,7 @@ test(`Respect prune setting in project manifest`, t => {
     'app.arc': arc,
     'public': {}
   })
-  sut(params, err => {
-    reset() // Must be reset before any tape tests are resolved because mock-fs#201
+  staticDeploy(t, err => {
     if (err) t.fail(err)
     t.ok(published.prune, 'Prune is enabled')
   })
@@ -200,8 +217,7 @@ test(`Respect folder setting in project manifest`, t => {
     'app.arc': arc,
     'some-folder': {}
   })
-  sut(params, err => {
-    reset() // Must be reset before any tape tests are resolved because mock-fs#201
+  staticDeploy(t, err => {
     if (err) t.fail(err)
     t.equal(published.folder, 'some-folder', 'Got correct folder setting')
   })
@@ -214,8 +230,7 @@ test(`Error if static folder isn't present`, t => {
   mockFs({
     'app.arc': arc
   })
-  sut(params, err => {
-    reset() // Must be reset before any tape tests are resolved because mock-fs#201
+  staticDeploy(t, err => {
     if (err) t.equal(err.message, '@static folder not found', 'Got error')
   })
 })
@@ -228,10 +243,8 @@ test(`Respect prefix param`, t => {
     'app.arc': arc,
     'public': {}
   })
-  let params = defaultParams()
   params.prefix = 'some-prefix'
-  sut(params, err => {
-    reset() // Must be reset before any tape tests are resolved because mock-fs#201
+  staticDeploy(t, err => {
     if (err) t.fail(err)
     t.equal(published.prefix, 'some-prefix', 'Prefix is unchanged')
   })
@@ -245,9 +258,14 @@ test(`Respect prefix setting in project manifest`, t => {
     'app.arc': arc,
     'public': {}
   })
-  sut(params, err => {
-    reset() // Must be reset before any tape tests are resolved because mock-fs#201
+  staticDeploy(t, err => {
     if (err) t.fail(err)
     t.equal(published.prefix, 'some-prefix', 'Got correct prefix setting')
   })
+})
+
+test('Teardown', t => {
+  t.plan(1)
+  awsMock.restore()
+  t.pass('Done')
 })